jout approvis et vente

+++ fenalait/backend/Fenalait/src/main/java/com/example/Fenalait/model/Approvissionnement.java
@@ -36,7 +36,7 @@ public class Approvissionnement extends  BaseEntity{
 	//@NotBlank(message = "Veuillez entrer la quantité d'approvissionnement !!")
 	private Double qteAppro;
 	
-	private long qteTotal;
+	private Double qteTotal;
 	
 	@Column(name="dateAppro")
 	//@Past(message = "La date ne peut être inférieure à la date courante !!")
